very meanful commit

+++ backend/app/src/main/java/com/congthanh/project/dto/ecommerce/CheckoutDTO.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -15,7 +17,7 @@ public class CheckoutDTO {
 
   private String customer;
 
-  private float total;
+  private BigDecimal total;
 
   private String address;
